backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.31.0",
    "cohere>=5.20.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.125.0",
    "openai-agents>=0.6.3",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
